OMAP: DSS2: Add zorder support on OMAP4
Add zorder Support on OMAP4, this feature allows deciding the visibility order
of the overlays based on the zorder value provided as a sysfs attribute of the
overlay object.

Since this feature is not supported on OMAP3 and OMAP2, the sysfs attibutes
always return zorder as 0 and do not write to dispc registers for non OMAP4
architectures.

Signed-off-by: Sumit Semwal <sumit.semwal@ti.com>
Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Mark Tyler <mark.tyler@ti.com>
Signed-off-by: Lajos Molnar <molnar@ti.com>